requirements.txt
# =============================================================================
# YOLO Training Pipeline - Requirements
# =============================================================================
# Install PyTorch first from: https://pytorch.org/get-started/
# Then: pip install -r requirements.txt
# =============================================================================

# Core ML Framework
torch>=2.9.0
torchvision>=0.24.0
ultralytics>=8.4.0

# Data Processing
numpy>=2.0.0
opencv-python>=4.10.0
pillow>=12.0.0
scipy>=1.14.0

# Visualization & Monitoring
matplotlib>=3.9.0
tensorboard>=2.18.0
seaborn>=0.13.0

# Utilities
pyyaml>=6.0
psutil>=5.9.0
tqdm>=4.66.0
pandas>=2.2.0

# Export formats (optional)
onnx>=1.16.0
onnxruntime>=1.18.0

# Development (optional)
# pytest>=8.0.0
# black>=24.0.0
# flake8>=7.0.0
